backend/upload_service/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-docx==0.8.11
Werkzeug==2.3.7
python-dotenv==1.0.0
